% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopVocaManager-package.R
\docType{package}
\name{OmopVocaManager-package}
\alias{OmopVocaManager}
\alias{OmopVocaManager-package}
\title{OmopVocaManager: Athena omop vocabulary export tool.}
\description{
This packager for uploading vocabulary files to DBMS. Download athena vocabulary files and run this function.
}
\keyword{internal}
