% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voca_manage.R
\docType{package}
\name{OmopVocaManager}
\alias{OmopVocaManager}
\alias{OmopVocaManager-package}
\title{Upload vocabulary files to DBMS}
\usage{
OmopVocaManager(connectionDetails, oracleTempSchema = NULL,
  vocabularyDatabaseSchema, dropIfExists = FALSE)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{vocabularyDatabaseSchema}{Schema name where your vocabulary in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'vocabulary_data.dbo'.}

\item{createTable}{Logical. If TRUE, drop existed tables and create new tables.}
}
\description{
This is start of OmopVocaManager. It needs DB connection information
}
\details{
OmopVocaManager

This function for help to upload vocabulary data to DB.
}
\examples{
\\donotrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

OmopVocaManager(connectionDetails = connectionDetails,
                oracleTempSchema = NULL,
                vocabularyDatabaseSchema = 'CDM_voca',
                dropIfExists = FALSE)
}

}
